Divide CLI commands across files
